feat(cli): implement compareOlaresVersion function and related tests
This commit introduces the compareOlaresVersion function, which mirrors the JavaScript implementation for version comparison, including handling of prerelease qualifiers. The function is designed to ensure consistent behavior across the CLI and backend. Additionally, comprehensive unit tests for this function have been added to validate its correctness against various versioning scenarios. Unused version comparison utilities have been removed to streamline the codebase.
